requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3
six==1.16.0
tornado==6.1
tzlocal==5.0.1
tinydb
responses
flask
requests
urllib3==1.26.15